requirements.txt
paho-mqtt
pandas
numpy
joblib
xgboost
streamlit
plotly
scikit-learn
fpdf2
python-dotenv
